+++ src/advin.h
@@ -28,20 +28,18 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+unsigned long long int get_long_long_uint(char *prompt);
+unsigned long int get_long_uint(char *prompt);
 unsigned int get_uint(char *prompt);
-
-short get_short(char *prompt);
-
-int get_int(char *prompt);
-
-long get_long(char *prompt);
+short unsigned int get_short_uint(char *prompt);
 
 long long get_long_long(char *prompt);
+long get_long(char *prompt);
+int get_int(char *prompt);
+short get_short(char *prompt);
 
 float get_float(char *prompt);
-
 double get_double(char *prompt);
-
 long double get_long_double(char *prompt);
 
 char get_char(char *prompt);