+++ test/get_long_test.c
@@ -1,0 +1,10 @@
+#include "../src/advin.h"
+#include <stdio.h>
+
+int main(void)
+{
+	printf("------------- TESTING -------------\n");
+	long int input = get_long("Number: ");
+
+	printf("Number: %ld\n", input);
+}